Prepare is no longer a pre-requisite for Compile ...
The makefiles here are meant to be invoked from within Java code.
And in the Java code, separate calls are made for both the targets.

Having prepare as a pre-requisite of compile only complicates things.
For one, if prepare makes the links as intended and compile tries to make
them again, then the make will error out. We could suppress the errors.
But that is not something I like to do