Brought back \insertQR into the template file ..
However, \insertQR is smarter now. If nothing is passed to it, then it
will not try and create a QR code - even if \printanswersfalse

The two changes together should make sub-part handling easier.

For standalone questions, only the first \insertQR - the one right after \question
would get a non-empty argument

For multi-part questions, every \insertQR other than the first one would get
a non-empty argument

The previous way of inserting \insertQR on the fly was upsetting the final
typesetting