Template file for generating complex curves using bc and fig4tex added
Here is how this template file will be used when complex curves need to be drawn
  1. The plot() and end_drawing() methods will be edited
  2. bc -l figure.bc > figure.tex command issued
  3. Copy-paste contents of figure.tex at the appropriate place in question.tex

Yes, (3) implies manual intervention. But still, using bc generates a lot of the
fig4tex code that one would otherwise have to write manually. Compared to that,
copying and pasting is a piece of cake