Added figure.bc to all previous questions in vault
But the aim really is to remove an inefficiency in the makefiles.
The inefficiency - in individual.mk - is coming because of the phony
target 'bc2fig' which should have a dependency on figure.bc but does
not because figure.bc was not always required.

Now, that phony target can be replaced with a real target with
real dependencies