Removed inefficiency in makefile due to phony target
The phony target was always firing because it didn't have any dependencies.
It didn't have any dependency because the dependency was a file (figure.bc)
that was not required to be present in every folder.

But now that it is, we can get replace the phony target w/ a real target w/
real dependencies