[Variable questions]: New commands and counters to support on-the-fly ..
.. creation of versions of the same question

Use only the commands expressly defined for this purpose in questions that you want
to make variable.

The idea is to present a slightly different version of the question to each
student so that plagiarism/cheating are harder to do. Then, hopefully, they
will put in honest effort even when they are not being monitored - like, when
it is a homework assignment