Ping'ing with the chapter_id extracted from source.xml
There is still some discrepancy in Quill on how it
handles new Question creation and how it handles
edits to old ones.

In particular, it does not seem to update the chapter_id
correctly for the latter when skills are set for it.

But it writes it to the XML. Hence, we will extract
and send the chapter ID from this script (as part of the make)